+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 All notable changes to this project will be documented in this file[^1].
 
 ## [Unreleased]
+### Fixed
+- multiple item work/object types
+
 ## [2.37.0] - 2021-12-03
 ### Added
 - newsletter sign-up form to info page